chore: "core" submodule from .gitmodules
The `.gitmodules` file was updated to remove the configuration for the "core" submodule. This includes deleting the `path` and `url` entries for the submodule, which was previously located at `core-` and linked to `https://github.com/sercanio/Myrtus.Clarity.Core.git`.

This change indicates that the repository no longer depends on or includes the "core" submodule, potentially simplifying the repository structure or reflecting that the submodule is no longer needed.